update incidence_by_time.py and test_incidence_by_time.py
